--- test/langtools/tools/javac/valhalla/primitive-classes/ValueAnnotationOnAnonymousClass.java
@@ -1,15 +1,0 @@
-/*
- * @test /nodynamiccopyright/
- * @summary Check valueness via __value__ annotation
- * @bug 8221699
- * @compile/fail/ref=ValueAnnotationOnAnonymousClass.out -XDrawDiagnostics -XDenablePrimitiveClasses ValueAnnotationOnAnonymousClass.java
- */
-
-class ValueAnnotationOnAnonymousClass {
-    interface I {}
-    primitive
-    public static void main(String args []) {
-        new primitive I() {
-        };
-    }
-}
