--- test/langtools/tools/javac/valhalla/primitive-classes/ValueAnnotationOnAnonymousClass.out
@@ -1,2 +1,0 @@
-ValueAnnotationOnAnonymousClass.java:11:24: compiler.err.mod.not.allowed.here: primitive
-1 error
