--- test/langtools/tools/javac/valhalla/primitive-classes/ValueAnnotationTest.out
@@ -1,3 +1,0 @@
-ValueAnnotationTest.java:25:1: compiler.err.illegal.combination.of.modifiers: interface, primitive
-ValueAnnotationTest.java:20:9: compiler.err.cant.assign.val.to.final.var: x
-2 errors
