--- test/langtools/tools/javac/valhalla/primitive-classes/ValueOverGenericsTest.java
@@ -1,19 +1,0 @@
-/*
- * @test /nodynamiccopyright/
- * @summary Value types cannot parameterize generic types (except under experimental mode)
- * @compile/fail/ref=ValueOverGenericsTest.out -XDrawDiagnostics -XDenablePrimitiveClasses ValueOverGenericsTest.java
- *
- */
-
-import java.util.ArrayList;
-import java.io.Serializable;
-
-primitive class ValueOverGenericsTest {
-    int x = 10;
-    ArrayList<ValueOverGenericsTest> ax = null;
-    void foo(ArrayList<? extends ValueOverGenericsTest> p) {
-        new <ValueOverGenericsTest> ArrayList<Object>();
-        this.<ValueOverGenericsTest>foo(null);
-        Object o = (ValueOverGenericsTest & Serializable) null;
-    }
-}
